Add `returning_on_create` to Persistence
Add ability for applications to specify a list of attributes that will be reloaded with a SQL `RETURNING` clause when new records are inserted.

Also handles a minor edge case if application code attempts to write to a virtual column. The values from `RETURNING` will always be used to update the instance's attributes.